task2

+++ 0x07-pointers_arrays_strings/2-strchr.c
@@ -10,7 +10,7 @@
 
 char *_strchr(char *s, char c)
 {
-	int i = 0;
+	unsigned long int i = 0;
 
 	if (s == NULL)
 		return (NULL);
